// backend/src/interfaces/student.interface.ts
export interface StudentRegisterInput {
  fullName: string;
  email: string;
  phone: string;
  password: string;
  confirmPassword: string;
}

export interface StudentDBInput {
  fullName: string;
  email: string;
  phone: string;
  password: string;
  isBlocked?: boolean;
}

export interface StudentOtp extends StudentRegisterInput{
  otp?: string;
  otpExpiresAt?: Date;
  isVerified?: boolean;
}

export interface ParentInfo {
     name: string;
     email: string;
     phone:string;
}

export interface contactInfo {
     parentInfo: ParentInfo;
     address: string;
     country: string;
     postalCode: string;
}

export interface AcademicDetails {
     institutionName: string;
     grade: string;
     syllabus: string;
}

export interface StudentProfile {
     _id: string;
     fullName: string;
     email: string;
     phone: string;
     password: string;
     age: number;
     gender: string;
     dateOfBirth: Date;
     contactInfo: contactInfo;
     academicDetails: AcademicDetails;
     profileImage: string;
     goal?: string;
     isVerified?: boolean;
     isBlocked: boolean;
     createdAt?: Date;
     updatedAt?: Date;
}